fix panel audit findings
