+++ panel/src/lib/types.ts
@@ -25,8 +25,8 @@ export type AgentSlug = KnownAgent | (string & {});
 // Back-compat alias for older call sites that still import AgentKind.
 export type AgentKind = AgentSlug;
 
-export type Ownership = "managed" | "observed" | "external";
-export type ProjectionMethod = "symlink" | "copy" | "materialize";
+export type Ownership = "managed" | "observed" | "external" | "unknown";
+export type ProjectionMethod = "symlink" | "copy" | "materialize" | "unknown";
 export type OpStatus = "ok" | "pending" | "err";
 export type SkillSourceStatus = "present" | "missing" | "non-compliant";
